backend/requirements.txt
fastapi
uvicorn
pillow
tensorflow
matplotlib
python-multipart
requests
